+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/bean/entity/HibernateProviderInfo.java
@@ -1,0 +1,102 @@
+package com.dwarfeng.judge.impl.bean.entity;
+
+import com.dwarfeng.judge.sdk.util.Constraints;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+
+import javax.persistence.*;
+import java.util.Optional;
+
+@Entity
+@IdClass(HibernateLongIdKey.class)
+@Table(name = "tbl_provider_info")
+public class HibernateProviderInfo implements Bean {
+
+    private static final long serialVersionUID = 2719198532386312830L;
+
+    // -----------------------------------------------------------主键-----------------------------------------------------------
+
+    @Id
+    @Column(name = "id", nullable = false, unique = true)
+    private Long longId;
+
+    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+
+    @Column(name = "enabled", nullable = false)
+    private boolean enabled;
+
+    @Column(name = "type", length = Constraints.LENGTH_TYPE, nullable = false)
+    private String type;
+
+    @Column(name = "param", columnDefinition = "TEXT")
+    private String param;
+
+    @Column(name = "remark", length = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    public HibernateProviderInfo() {
+    }
+
+
+    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+
+    public HibernateLongIdKey getKey() {
+        return Optional.ofNullable(longId).map(HibernateLongIdKey::new).orElse(null);
+    }
+
+    public void setKey(HibernateLongIdKey idKey) {
+        this.longId = Optional.ofNullable(idKey).map(HibernateLongIdKey::getLongId).orElse(null);
+    }
+
+    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+
+    public Long getLongId() {
+        return longId;
+    }
+
+    public void setLongId(Long longId) {
+        this.longId = longId;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getParam() {
+        return param;
+    }
+
+    public void setParam(String param) {
+        this.param = param;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + "(" +
+                "longId = " + longId + ", " +
+                "enabled = " + enabled + ", " +
+                "type = " + type + ", " +
+                "param = " + param + ", " +
+                "remark = " + remark + ")";
+    }
+}