+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/cache/ProviderSupportCacheImpl.java
@@ -1,0 +1,95 @@
+package com.dwarfeng.judge.impl.cache;
+
+import com.dwarfeng.judge.sdk.bean.entity.FastJsonProviderSupport;
+import com.dwarfeng.judge.stack.bean.entity.ProviderSupport;
+import com.dwarfeng.judge.stack.cache.ProviderSupportCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.CacheException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class ProviderSupportCacheImpl implements ProviderSupportCache {
+
+    private final RedisBatchBaseCache<StringIdKey, ProviderSupport, FastJsonProviderSupport> batchBaseCache;
+
+    public ProviderSupportCacheImpl(RedisBatchBaseCache<StringIdKey, ProviderSupport, FastJsonProviderSupport> batchBaseCache) {
+        this.batchBaseCache = batchBaseCache;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(StringIdKey key) throws CacheException {
+        return batchBaseCache.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public ProviderSupport get(StringIdKey key) throws CacheException {
+        return batchBaseCache.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void push(ProviderSupport value, long timeout) throws CacheException {
+        batchBaseCache.push(value, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(StringIdKey key) throws CacheException {
+        batchBaseCache.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void clear() throws CacheException {
+        batchBaseCache.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return batchBaseCache.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return batchBaseCache.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<ProviderSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return batchBaseCache.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<ProviderSupport> entities, long timeout) throws CacheException {
+        batchBaseCache.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        batchBaseCache.batchDelete(keys);
+    }
+}