+++ judge-sdk/src/main/java/com/dwarfeng/judge/sdk/bean/entity/WebInputProviderInfo.java
@@ -1,0 +1,108 @@
+package com.dwarfeng.judge.sdk.bean.entity;
+
+import com.alibaba.fastjson.annotation.JSONField;
+import com.dwarfeng.judge.sdk.util.Constraints;
+import com.dwarfeng.judge.stack.bean.entity.ProviderInfo;
+import com.dwarfeng.subgrade.sdk.bean.key.WebInputLongIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+import org.hibernate.validator.constraints.Length;
+
+import javax.validation.Valid;
+import java.util.Objects;
+
+/**
+ * WebInput 提供器信息。
+ *
+ * @author wangyc
+ * @since 2.1.0
+ */
+public class WebInputProviderInfo implements Bean {
+
+    private static final long serialVersionUID = -6086744238699864316L;
+
+    public static ProviderInfo toStackBean(WebInputProviderInfo webInput) {
+        if (Objects.isNull(webInput)) {
+            return null;
+        } else {
+            return new ProviderInfo(
+                    WebInputLongIdKey.toStackBean(webInput.getKey()),
+                    webInput.isEnabled(),
+                    webInput.getType(),
+                    webInput.getParam(),
+                    webInput.getRemark()
+            );
+        }
+    }
+
+    @JSONField(name = "key")
+    @Valid
+    private WebInputLongIdKey key;
+
+    @JSONField(name = "enabled")
+    private boolean enabled;
+
+    @JSONField(name = "type")
+    private String type;
+
+    @JSONField(name = "param")
+    private String param;
+
+    @JSONField(name = "remark")
+    @Length(max = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    public WebInputProviderInfo() {
+    }
+
+    public WebInputLongIdKey getKey() {
+        return key;
+    }
+
+    public void setKey(WebInputLongIdKey key) {
+        this.key = key;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getParam() {
+        return param;
+    }
+
+    public void setParam(String param) {
+        this.param = param;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return "WebInputProviderInfo{" +
+                "key=" + key +
+                ", enabled=" + enabled +
+                ", type=" + type + '\'' +
+                ", param=" + param + '\'' +
+                ", remark=" + remark + '\'' +
+                '}';
+    }
+}
+