+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/cache/ProviderInfoCache.java
@@ -1,0 +1,14 @@
+package com.dwarfeng.judge.stack.cache;
+
+import com.dwarfeng.judge.stack.bean.entity.ProviderInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
+
+/**
+ * 提供器信息缓存。
+ *
+ * @author wangyc
+ * @since 2.1.0
+ */
+public interface ProviderInfoCache extends BatchBaseCache<LongIdKey, ProviderInfo> {
+}