+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/service/ProviderSupportMaintainService.java
@@ -1,0 +1,18 @@
+package com.dwarfeng.judge.stack.service;
+
+import com.dwarfeng.judge.stack.bean.entity.ProviderSupport;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.service.BatchCrudService;
+import com.dwarfeng.subgrade.stack.service.EntireLookupService;
+import com.dwarfeng.subgrade.stack.service.PresetLookupService;
+
+/**
+ * 提供器支持维护服务。
+ *
+ * @author wangyc
+ * @since 2.1.0
+ */
+public interface ProviderSupportMaintainService extends BatchCrudService<StringIdKey, ProviderSupport>,
+        EntireLookupService<ProviderSupport>, PresetLookupService<ProviderSupport> {
+
+}